2.5.0 beta 2 to Master branch

+++ scripts/MPBuildServer.sh
@@ -2,7 +2,7 @@
 #
 # -------------------------------------------------------------
 # Script: MPBuildServer.sh
-# Version: 1.0
+# Version: 1.2
 #
 # Description:
 # This is a very simple script to demonstrate how to automate
@@ -16,7 +16,49 @@ MPBASE="/Library/MacPatch"
 MPSERVERBASE="/Library/MacPatch/Server"
 GITROOT="/Library/MacPatch/tmp/MacPatch"
 BUILDROOT="/Library/MacPatch/tmp/build/Server"
+SRC_DIR="${MPSERVERBASE}/conf/src"
 TCATSRV=0
+J2EE_SW="apache-tomcat-7.0.55.tar.gz"
+
+XOSTYPE=`uname -s`
+USELINUX=false
+USEMACOS=false
+OWNERGRP="79:70"
+
+if [ "`whoami`" != "root" ] ; then   # If not root user,
+   # Run this script again as root
+   echo
+   echo "You must be an admin user to run this script."
+   echo "Please re-run the script using sudo."
+   echo
+   exit 1;
+fi
+
+# -----------------------------------
+# OS Check
+# -----------------------------------
+
+# Check and set os type
+if [ $XOSTYPE == "Linux" ]; then
+	USELINUX=true
+	OWNERGRP="www-data:www-data"
+	getent passwd www-data > /dev/null 2&>1
+	if [ $? -eq 0 ]; then
+		echo "www-data user exists"
+	else
+    	echo "Create user www-data"
+		useradd -r -M -s /dev/null -U www-data
+	fi
+elif [ $XOSTYPE == "Darwin" ]; then
+	USEMACOS=true
+else
+  	echo "OS Type $XOSTYPE is not supported. Now exiting."
+  	exit 1; 
+fi
+
+# -----------------------------------
+# Main
+# -----------------------------------
 
 if [ -d "$BUILDROOT" ]; then
 	rm -rf ${BUILDROOT}
@@ -36,8 +78,8 @@ fi
 # Tomcat support is new and going to replace Jetty
 # ------------------------------------
 
-read -p "Use experimental Tomcat config [N]: " TCATSRV
-TCATSRV=${TCATSRV:-0}
+read -p "Use older Jetty J2EE Server [N]: " TCATSRV
+TCATSRV=${TCATSRV:-1}
 
 # ------------------
 # Create Skeleton Dir Structure
@@ -49,26 +91,46 @@ mkdir -p /Library/MacPatch/Content/Web/clients
 mkdir -p /Library/MacPatch/Content/Web/patches
 mkdir -p /Library/MacPatch/Content/Web/sav
 mkdir -p /Library/MacPatch/Content/Web/sw
+mkdir -p /Library/MacPatch/Content/Web/tools
 mkdir -p /Library/MacPatch/Server
 mkdir -p /Library/MacPatch/Server/lib
 mkdir -p /Library/MacPatch/Server/Logs
 
 # ------------------
-# Compile the agent components
-# ------------------
-xcodebuild clean build -project ${GITROOT}/MacPatch/MacPatch.xcodeproj -target SERVER_BUILD SYMROOT=${BUILDROOT}
-
-# ------------------
-# Remove the build and symbol files
-# ------------------
-find ${BUILDROOT} -name "*.build" -print | xargs -I{} rm -rf {}
-find ${BUILDROOT} -name "*.dSYM" -print | xargs -I{} rm -rf {}
-
-# ------------------
 # Copy compiled files
 # ------------------
 cp -R ${GITROOT}/MacPatch\ Server/Server ${MPBASE}
-cp -R ${BUILDROOT}/Release/ ${MPSERVERBASE}/bin
+cp ${MPSERVERBASE}/conf/Content/Web/tools/MPAgentUploader.app.zip /Library/MacPatch/Content/Web/tools/
+
+if $USEMACOS; then
+	# ------------------
+	# Compile the agent components
+	# ------------------
+	xcodebuild clean build -project ${GITROOT}/MacPatch/MacPatch.xcodeproj -target SERVER_BUILD SYMROOT=${BUILDROOT}
+
+	# ------------------
+	# Remove the build and symbol files
+	# ------------------
+	find ${BUILDROOT} -name "*.build" -print | xargs -I{} rm -rf {}
+	find ${BUILDROOT} -name "*.dSYM" -print | xargs -I{} rm -rf {}
+
+	# ------------------
+	# Copy compiled files
+	# ------------------
+	cp -R ${BUILDROOT}/Release/ ${MPSERVERBASE}/bin
+fi
+
+# ------------------
+# Setup Tomcat
+# ------------------
+
+if [ $TCATSRV == 1 ]; then
+	mkdir -p "${MPSERVERBASE}/apache-tomcat"
+	tar xvfz ${SRC_DIR}/${J2EE_SW} --strip 1 -C ${MPSERVERBASE}/apache-tomcat
+	chmod +x ${MPSERVERBASE}/apache-tomcat/bin/*
+	rm -rf ${MPSERVERBASE}/apache-tomcat/webapps/docs
+	rm -rf ${MPSERVERBASE}/apache-tomcat/webapps/examples
+fi
 
 # ------------------
 # Build Apache
@@ -79,22 +141,26 @@ ${MPSERVERBASE}/conf/scripts/MPHttpServerBuild.sh
 # Link & Set Permissions
 # ------------------
 ln -s ${MPSERVERBASE}/conf/Content/Doc ${MPBASE}/Content/Doc
-chown -R root:admin ${MPSERVERBASE}
+chown -R $OWNERGRP ${MPSERVERBASE}
+
 if [ $TCATSRV == 0 ]; then
 	rm -rf "${MPSERVERBASE}/apache-tomcat"
 	chmod -R 0775 ${MPSERVERBASE}/jetty-mpsite
-	chown -R 79:70 ${MPSERVERBASE}/jetty-mpsite
+	chown -R $OWNERGRP ${MPSERVERBASE}/jetty-mpsite
 	chmod -R 0775 ${MPSERVERBASE}/jetty-mpwsl
-	chown -R 79:70 ${MPSERVERBASE}/jetty-mpwsl
+	chown -R $OWNERGRP ${MPSERVERBASE}/jetty-mpwsl
 else
 	cp -r ${MPSERVERBASE}/apache-tomcat ${MPSERVERBASE}/tomcat-mpws
 	mv ${MPSERVERBASE}/apache-tomcat ${MPSERVERBASE}/tomcat-mpsite
-
+	mkdir -p ${MPSERVERBASE}/tomcat-mpws/InvData/Files
+	mkdir -p ${MPSERVERBASE}/tomcat-mpws/InvData/Errors
+	mkdir -p ${MPSERVERBASE}/tomcat-mpws/InvData/Processed
+	
+	cp -r "${MPSERVERBASE}/conf/lib/systemcommand.jar" "${MPSERVERBASE}/jetty-mpwsl/webapps/mpwsl/WEB-INF/lib/systemcommand.jar"
 	chmod -R 0775 "${MPSERVERBASE}/jetty-mpwsl/webapps/mpwsl"
-	chown -R 79:70 "${MPSERVERBASE}/jetty-mpwsl/webapps/mpwsl"
+	chown -R $OWNERGRP "${MPSERVERBASE}/jetty-mpwsl/webapps/mpwsl"
 	rm -rf  "${MPSERVERBASE}/tomcat-mpws/webapps/ROOT"
 	jar cf "${MPSERVERBASE}/conf/tomcat/mpws/ROOT.war" -C "${MPSERVERBASE}/jetty-mpwsl/webapps/mpwsl" .
-	#cp -r "${MPSERVERBASE}/jetty-mpwsl/webapps/mpwsl" "${MPSERVERBASE}/tomcat-mpws/webapps/ROOT"
 	cp "${MPSERVERBASE}/conf/tomcat/mpws/ROOT.war" "${MPSERVERBASE}/tomcat-mpws/webapps"
 	cp "${MPSERVERBASE}/conf/tomcat/mpws/bin/setenv.sh" "${MPSERVERBASE}/tomcat-mpws/bin/setenv.sh"
 	cp "${MPSERVERBASE}/conf/tomcat/mpws/bin/launchdTomcat.sh" "${MPSERVERBASE}/tomcat-mpws/bin/launchdTomcat.sh"
@@ -103,11 +169,11 @@ else
 	cp -r "${MPSERVERBASE}/conf/tomcat/mpws/conf/web.xml" "${MPSERVERBASE}/tomcat-mpws/conf/web.xml"
 	rm -rf "${MPSERVERBASE}/jetty-mpwsl"
 
+	cp -r "${MPSERVERBASE}/conf/lib/systemcommand.jar" "${MPSERVERBASE}/jetty-mpsite/webapps/mp/WEB-INF/lib/systemcommand.jar"
 	chmod -R 0775 "${MPSERVERBASE}/jetty-mpsite/webapps/mp"
-	chown -R 79:70 "${MPSERVERBASE}/jetty-mpsite/webapps/mp"
+	chown -R $OWNERGRP "${MPSERVERBASE}/jetty-mpsite/webapps/mp"
 	rm -rf "${MPSERVERBASE}/tomcat-mpsite/webapps/ROOT"
 	jar cf "${MPSERVERBASE}/conf/tomcat/mpsite/ROOT.war" -C "${MPSERVERBASE}/jetty-mpsite/webapps/mp" .
-	#cp -r "${MPSERVERBASE}/jetty-mpsite/webapps/mp" "${MPSERVERBASE}/tomcat-mpsite/webapps/ROOT"
 	cp "${MPSERVERBASE}/conf/tomcat/mpsite/ROOT.war" "${MPSERVERBASE}/tomcat-mpsite/webapps"
 	cp "${MPSERVERBASE}/conf/tomcat/mpsite/bin/setenv.sh" "${MPSERVERBASE}/tomcat-mpsite/bin/setenv.sh"
 	cp "${MPSERVERBASE}/conf/tomcat/mpsite/bin/launchdTomcat.sh" "${MPSERVERBASE}/tomcat-mpsite/bin/launchdTomcat.sh"
@@ -117,12 +183,12 @@ else
 	rm -rf "${MPSERVERBASE}/jetty-mpsite"
 
 	chmod -R 0775 ${MPSERVERBASE}/tomcat-mpws
-	chown -R 79:70 ${MPSERVERBASE}/tomcat-mpws
+	chown -R $OWNERGRP ${MPSERVERBASE}/tomcat-mpws
 	chmod -R 0775 ${MPSERVERBASE}/tomcat-mpsite
-	chown -R 79:70 ${MPSERVERBASE}/tomcat-mpsite
+	chown -R $OWNERGRP ${MPSERVERBASE}/tomcat-mpsite
 fi
 
-chown -R 79:70 ${MPSERVERBASE}/Logs
+chown -R $OWNERGRP ${MPSERVERBASE}/Logs
 chmod 0775 ${MPSERVERBASE}
 
 
@@ -134,4 +200,4 @@ find ${MPSERVERBASE} -name ".mpRM" -print | xargs -I{} rm -rf {}
 # ------------------
 # Create Archive
 # ------------------
-ditto -c -k ${MPSERVERBASE} ${MPBASE}/MacPatch_Server.zip
+zip -r ${MPBASE}/MacPatch_Server.zip ${MPSERVERBASE}
